+++ CHANGELOG.md
@@ -16,7 +16,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
  - Repeater changed to use configured callbacks instead of a dumb event
  - Scheduled action callbacks allow a TimeSpan to be returned to delay actions
- - Workaround for logged XML format exceptions when failing to parse floats
+
+## [13.0.0] - 2020-09-03
+### Added
+ - Added util methods for removing duplicate whitespace in strings
+ - Added dequeue overload to ScrollQueue
+
+### Changed
+ - Replaced Crestron Unzip with Yallie Unzip
+ - Fixed "version" regex for 4-series
+ - Fixed date parsing error for 4-series
+ - Split SimplSharpProMono runtime environment into SimplSharpProSever
+ - Fixed log formatting on 4-series
 
 ## [12.1.0] - 2020-07-14
 ### Added